guardforge/backend/requirements.txt
# GuardForge — Backend Dependencies
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.11,<3.0
pydantic-settings>=2.0,<3.0
sqlalchemy>=2.0,<3.0
aiosqlite>=0.20,<1.0
cryptography>=44.0,<45.0
PyYAML>=6.0,<7.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
httpx>=0.27,<1.0
